server/requirements.txt
flask
waitress
pycryptodome
